% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_trip.R
\name{add_tagon_time}
\alias{add_tagon_time}
\title{Add column for tag on time to tag off transactions}
\usage{
add_tagon_time(tr_df)
}
\arguments{
\item{tr_df}{dataframe of transactions}
}
\description{
Add column for tag on time to tag off transactions
}
