% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journeys.R
\name{add_transfer_time}
\alias{add_transfer_time}
\title{Adds column classifying transfers as given by the transfer rules dataframe}
\usage{
add_transfer_time(tr_df)
}
\arguments{
\item{tr_df}{a dataframe of rides}
}
\description{
Adds column classifying transfers as given by the transfer rules dataframe
}
