% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_trip.R
\name{add_trip_cost}
\alias{add_trip_cost}
\title{Add column for previous purse amount and actual trip cost for trips with subtypes 2/3}
\usage{
add_trip_cost(tr_df)
}
\arguments{
\item{tr_df}{dataframe of transactions}
}
\value{
tr_df dataframe of transactions
}
\description{
Add column for previous purse amount and actual trip cost for trips with subtypes 2/3
}
