% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{as_bart_journeys}
\alias{as_bart_journeys}
\title{Spreads multiple transactions across columns into one-row-per-bart-trip (with a focus on transfers in and out)}
\usage{
as_bart_journeys(tr_df)
}
\arguments{
\item{tr_df}{a sample of transactions, effectively from the raw sfofaretransactions table, joined to other tables}
}
\description{
Spreads multiple transactions across columns into one-row-per-bart-trip (with a focus on transfers in and out)
}
