% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_trip.R
\name{as_rides}
\alias{as_rides}
\title{Drop rows with tag on subtypes after recording the relevant tag on information in the tag off transactions}
\usage{
as_rides(tr_df)
}
\arguments{
\item{tr_df}{dataframe of transactions}
}
\value{
tr_df dataframe of transactions
}
\description{
Drop rows with tag on subtypes after recording the relevant tag on information in the tag off transactions
}
