% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{bart_identify}
\alias{bart_identify}
\title{Adds columns identifying bart transactions, including lags and leads}
\usage{
bart_identify(tr_df)
}
\arguments{
\item{tr_df}{a dataframe of transactions}
}
\description{
Adds columns identifying bart transactions, including lags and leads
}
