% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{bart_lag_and_lead_metadata}
\alias{bart_lag_and_lead_metadata}
\title{Adds columns describing the route, time, and operator name of previous and later transactions}
\usage{
bart_lag_and_lead_metadata(tr_df)
}
\arguments{
\item{tr_df}{a dataframe of transactions}
}
\value{
tr_df a dataframe of transactions
}
\description{
Adds columns describing the route, time, and operator name of previous and later transactions
}
