% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_transfers.R
\name{bart_to_bart}
\alias{bart_to_bart}
\title{Filters transactions to BART intra-transfers only}
\usage{
bart_to_bart(tr_df)
}
\arguments{
\item{tr_df}{a dataframe of transactions}
}
\value{
tr_df a dataframe of transactions
}
\description{
Filters transactions to BART intra-transfers only
}
