% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{bart_transactions_per_user}
\alias{bart_transactions_per_user}
\title{Adds a column with a counts of the number of bart transactions per user and diff with all transactions}
\usage{
bart_transactions_per_user(tr_df)
}
\arguments{
\item{tr_df}{a dataframe of transactions}
}
\value{
tr_df a dataframe of transactions with a bart_tr_count and tr_count_diff column
}
\description{
Adds a column with a counts of the number of bart transactions per user and diff with all transactions
}
