% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{Build_Database}
\alias{Build_Database}
\title{Build a database of transfers}
\usage{
Build_Database(df_row, data_dir, transfer_rules_df, sampling_rate_float,
  output_file_string, append_boolean)
}
\arguments{
\item{df_row}{a row from transfer_filenames_df, which is included in the package}

\item{data_dir}{the directory in which the transfer and transaction data resides}

\item{transfer_rules_df}{a set of travel model rules on what constituted a transfer}
}
\description{
The anonymous Clipper data is potentially an excellent source for quantifying operator-to-operator transit movements.
Here, we seek to understand transfers in the travel model sense,
i.e. a movement between an origin and destination that requires
a tranfer between or within transit operators. This definition differs from transfers
as defined by certain transit agencies, which generally allow multiple boardings at no or reduced cost
during a narrow time window.
Here, we implement a set of rules to build a database of transfers.
This script uses data derived from the anonymous Clipper data
see Extract Transfer for Random Weekday to CSV.R.
}
