% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggt_transfers.R
\name{combined_ggt_transactions}
\alias{combined_ggt_transactions}
\title{Creates combined dataframe of multiple transactions per row}
\usage{
combined_ggt_transactions(tr_df)
}
\arguments{
\item{tr_df}{a dataframe of transactions}
}
\description{
Creates combined dataframe of multiple transactions per row
}
