% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_df.R
\name{create_transfer_df}
\alias{create_transfer_df}
\title{Creates a summary table of all combinations of transfers}
\usage{
create_transfer_df(tr_df, mins)
}
\arguments{
\item{tr_df}{a dataframe of all transactions}

\item{mins}{time period to qualify as a transfer (in minutes)}
}
\description{
Creates a summary table of all combinations of transfers
}
