% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{day_of_transactions}
\alias{day_of_transactions}
\title{Pull all transactions for a given number of users}
\usage{
day_of_transactions(rs, date1, drop_existing_table = FALSE)
}
\arguments{
\item{rs}{RPostgres connection to database}

\item{date1}{date to query}

\item{drop_existing_table}{whether to drop an existing partitioned table for this date}
}
\value{
tibble a tibble with the results for the day from the db
}
\description{
Pull all transactions for a given number of users
}
