% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{descriptive_tables}
\alias{descriptive_tables}
\title{Get descriptive tables for transactions}
\usage{
descriptive_tables()
}
\value{
a list of dbplyr/dplyr tbl tables: operators, routes, locations
}
\description{
Results can be joined with transactions to make them more legible.
Expects database to be configured as for connect_rs()
}
