% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{devices_for_day}
\alias{devices_for_day}
\title{Get the device transactions for a day
This function has been deprecated and should be used with care}
\usage{
devices_for_day(partition_time = "10:00:00", start_date = "2016-01-01",
  drop_existing_table = FALSE)
}
\description{
Get the device transactions for a day
This function has been deprecated and should be used with care
}
