% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_sequence_by_date}
\alias{extract_sequence_by_date}
\title{importFrom odbc dbClearResult dbDisconnect dbSendQuery}
\usage{
extract_sequence_by_date(con)
}
\description{
importFrom odbc dbClearResult dbDisconnect dbSendQuery
}
