% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{fares_for_day}
\alias{fares_for_day}
\title{Partition the transaction table by travel survey/model day (starting at 3 am)}
\usage{
fares_for_day(partition_time = "10:00:00", start_date = "2016-01-01",
  drop_existing_table = FALSE)
}
\arguments{
\item{partition_time}{time to use as the start of a "day." defaults to 3 AM PST (10 AM UTC)}

\item{start_date}{date this is the start date for the day (starting at 3 am at the partition time above). note that the day may "end" at 3 am the next day.}

\item{drop_existing_table}{whether to drop an existing partitioned table for this date}
}
\description{
Creates a table for the day in the db on the clipper_day schema and returns the name of it.
Using in-database functions, returns card_id as a SHA_1 hash and time zone to PST.
}
