% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{get_product_description}
\alias{get_product_description}
\title{Add human-readable product names to transaction entries}
\usage{
get_product_description(tr_tbl)
}
\arguments{
\item{tr_tbl}{a transactions tbl connection as returned from, for example day_of_transactions(rs,date)}
}
\description{
Add human-readable product names to transaction entries
}
\examples{
transactions_tbl <- day_of_transactions(rs,date)
transactions_tbl_with_product_names <- get_product_description(transactions_tbl)
prod_summary <- transactions_tbl2 \%>\%
 group_by(product_description) \%>\%
 summarise(count=n_distinct(cardid_anony))

}
