% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggt_transfers.R
\name{ggt_identify}
\alias{ggt_identify}
\title{Adds columns identifying bart transactions, including lags and leads}
\usage{
ggt_identify(tr_df)
}
\arguments{
\item{tr_df}{a dataframe of transactions}
}
\description{
Adds columns identifying bart transactions, including lags and leads
}
