% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggt_transfers.R
\name{ggt_to_ggt}
\alias{ggt_to_ggt}
\title{Filters transactions to ggt intra-transfers only}
\usage{
ggt_to_ggt(tr_df)
}
\arguments{
\item{tr_df}{a dataframe of transactions}
}
\description{
Filters transactions to ggt intra-transfers only
}
