% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_df.R
\name{identify_transfer_for_time}
\alias{identify_transfer_for_time}
\title{Adds column to indicate a transfer under given time windows}
\usage{
identify_transfer_for_time(tr_df, mins)
}
\arguments{
\item{tr_df}{a dataframe of transactions}

\item{mins}{time period to qualify as a transfer (in minutes)}
}
\description{
Adds column to indicate a transfer under given time windows
}
