% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{sample_a_day_with_devices}
\alias{sample_a_day_with_devices}
\title{Sample transactions in a day with corresponding device table
This function has been deprecated and should be used with care}
\usage{
sample_a_day_with_devices(rs, date1, users)
}
\description{
Sample transactions in a day with corresponding device table
This function has been deprecated and should be used with care
}
