% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{sample_day_of_transactions}
\alias{sample_day_of_transactions}
\title{Sample a day of transactions for a given number of users}
\usage{
sample_day_of_transactions(rs, date1, n_users, drop_existing_table = FALSE)
}
\arguments{
\item{rs}{RPostgres connection to database}

\item{date1}{date to query}

\item{n_users}{number of users}

\item{drop_existing_table}{whether to drop an existing partitioned table for this date}
}
\value{
tbl a dbplyr/dplyr tbl table
}
\description{
Sample a day of transactions for a given number of users
}
