% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{spread_time_column}
\alias{spread_time_column}
\title{get a nicely formatted time dataframe
(yday, hour, yday, row, etc) for a column
we do the date parsing this way because of the UTC/datetime crossover}
\usage{
spread_time_column(timestamp_col, prefix = "t_")
}
\description{
get a nicely formatted time dataframe
(yday, hour, yday, row, etc) for a column
we do the date parsing this way because of the UTC/datetime crossover
}
