% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{transactions_per_user}
\alias{transactions_per_user}
\title{Adds a column with a counts of the number of transactions per user}
\usage{
transactions_per_user(tr_df)
}
\arguments{
\item{tr_df}{a dataframe of transactions}
}
\description{
Adds a column with a counts of the number of transactions per user
}
