% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_filenames.R
\docType{data}
\name{transfer_filenames_df}
\alias{transfer_filenames_df}
\title{Metadata on Input filenames for Transfers and Transactions}
\description{
A data frame of filenames and metadata for anonymized transfers and transactions partitioned by month
}
\details{
Data obtained from MTCSHARE/Data/Clipper
}
\seealso{
Build_Database
}
