% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journeys.R
\name{add_journey_id}
\alias{add_journey_id}
\title{Adds column grouping applicable rides as numbered journeys}
\usage{
add_journey_id(tr_df)
}
\arguments{
\item{tr_df}{a dataframe of rides}
}
\value{
tr_df a dataframe of rides
}
\description{
Adds column grouping applicable rides as numbered journeys
}
