% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_trip.R
\name{drop_tagons}
\alias{drop_tagons}
\title{Drop rows with tag on subtypes after recording the relevant tag on information in the tag off transactions}
\usage{
drop_tagons(tr_df)
}
\arguments{
\item{tr_df}{dataframe of transactions}
}
\description{
Drop rows with tag on subtypes after recording the relevant tag on information in the tag off transactions
}
