% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{get_product_description}
\alias{get_product_description}
\title{Add human-readable product names to transaction entries}
\usage{
get_product_description(tr_tbl)
}
\arguments{
\item{tr_tbl}{a transactions tibble connection as returned from, for example day_of_transactions(rs,date)}
}
\value{
tr_tbl a transactions tibble with a text value based product description column
}
\description{
Add human-readable product names to transaction entries
}
